requirements.txt
Django==4.2.21
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.5
Pillow==10.1.0
PyJWT==2.8.0
gunicorn==21.2.0
whitenoise==6.6.0 